requirements.txt
# Core ML Libraries (using versions with pre-built Windows wheels)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning / ML Extensions
mlxtend>=0.22.0

# MLOps & Experiment Tracking
mlflow>=2.9.0
dvc[all]>=3.30.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Data Validation & Quality
great-expectations>=0.18.0
evidently>=0.4.0

# Model Serialization
joblib>=1.3.0
cloudpickle>=2.2.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
requests>=2.31.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Database (for MLflow)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Cloud Storage Support (optional but professional)
boto3>=1.34.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.14.0

# Security
cryptography>=41.0.0
